FLINK-2740 Addressing feedback - moving example topologies to examples package, providing configurable wait time, protecting against NPE on transaction
