+++ flink-staging/flink-streaming/flink-streaming-connectors/flink-connector-nifi/src/main/java/org/apache/flink/streaming/connectors/nifi/examples/NiFiSourceTopologyExample.java
@@ -14,12 +14,14 @@
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
-package org.apache.flink.streaming.connectors.nifi;
+package org.apache.flink.streaming.connectors.nifi.examples;
 
 import org.apache.flink.api.common.functions.MapFunction;
 import org.apache.flink.streaming.api.datastream.DataStream;
 import org.apache.flink.streaming.api.environment.StreamExecutionEnvironment;
 import org.apache.flink.streaming.api.functions.source.SourceFunction;
+import org.apache.flink.streaming.connectors.nifi.NiFiDataPacket;
+import org.apache.flink.streaming.connectors.nifi.NiFiSource;
 import org.apache.nifi.remote.client.SiteToSiteClient;
 import org.apache.nifi.remote.client.SiteToSiteClientConfig;
 
@@ -36,10 +38,11 @@ public class NiFiSourceTopologyExample {
 		SiteToSiteClientConfig clientConfig = new SiteToSiteClient.Builder()
 				.url("http://localhost:8080/nifi")
 				.portName("Data for Flink")
+				.requestBatchCount(5)
 				.buildConfig();
 
 		SourceFunction<NiFiDataPacket> nifiSource = new NiFiSource(clientConfig);
-		DataStream<NiFiDataPacket> streamSource = env.addSource(nifiSource); //.setParallelism(2);
+		DataStream<NiFiDataPacket> streamSource = env.addSource(nifiSource).setParallelism(2);
 
 		DataStream<String> dataStream = streamSource.map(new MapFunction<NiFiDataPacket, String>() {
 			@Override
